Refactor and create modal context for mobile
Removed the refs & useImperativeHandle from App component & MobileNav
component and replaced it with a modal context that can be imported
where needed. Cleaner and more readable code.
