+++ src/components/plan-page/Plan.jsx
@@ -2,8 +2,8 @@ import React, { Suspense, lazy } from 'react';
 import Hero from './hero-section/Hero';
 import PlanSteps from './plan-section/PlanSteps';
 import Options from './options-section/Options';
-import OrderModalProvider from './context/OrderModalContext';
-import StoreProvider from './context/Context';
+import OrderModalProvider from '../../context/OrderModalContext';
+import StoreProvider from '../../context/Context';
 
 const OrderModal = lazy(() => import('./order-modal/OrderModal'));
