+++ src/components/plan-page/options-section/OptionContainer.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useContext } from 'react';
-import { StoreContext } from '../context/Context';
+import { StoreContext } from '../../../context/Context';
 import styles from './_options.module.scss';
 
 const OptionContainer = ({ options }) => {
