+++ src/components/plan-page/order-modal/OrderModal.jsx
@@ -1,8 +1,8 @@
 import React, { useContext, useEffect, useRef } from 'react';
 import ReactDOM from 'react-dom';
 import { withRouter } from 'react-router-dom';
-import { StoreContext } from '../context/Context';
-import { OrderModalContext } from '../context/OrderModalContext';
+import { StoreContext } from '../../../context/Context';
+import { OrderModalContext } from '../../../context/OrderModalContext';
 import calculateTotal from './priceTotal';
 import styles from './_order-modal.module.scss';
